# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NowBuy API
servers:
  - url: /api
    description: Base API path
...
paths:
  /admin/users/{id}/reset-code:
    post:
      tags: [PasswordReset]
      operationId: createUserResetCode
      summary: Generate a one-time password reset code for a user (admin only).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Reset code generated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResetCodeResponse' }
...
  /auth/redeem-reset-code:
    post:
      tags: [PasswordReset]
      operationId: redeemResetCode
      summary: Redeem a password reset code and set a new password.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RedeemResetCodeRequest' }
      responses:
        '200':
          description: Password updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthUser' }
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin/pm only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [buyer, admin, pm] }
      required: [id, email, name, role]
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        currency: { type: string }
        imageUrl: { type: string }
        rating: { type: number }
        stock: { type: integer }
        sellerName: { type: string }
        tags:
          type: array
          items: { type: string }
      required: [id, name, description, category, price, currency, imageUrl, rating, stock, sellerName, tags]
    ResetCodeResponse:
      type: object
      required: [code, expiresAt, scope, userEmail]
      properties:
        code: { type: string }
        scope: { type: string, enum: [user, admin] }
        userEmail: { type: string }
        expiresAt: { type: string, format: date-time }
    RedeemResetCodeRequest:
      type: object
      required: [email, code, newPassword]
      properties:
        email: { type: string, format: email }
        code: { type: string, minLength: 4 }
        newPassword: { type: string, minLength: 6 }
...
    CreateProductRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        currency: { type: string, default: USD }
        imageUrl: { type: string, minLength: 1 }
        stock: { type: integer, minimum: 0 }
        sellerName: { type: string, minLength: 1 }
        tags:
          type: array
          items: { type: string }
      required: [name, description, category, price, imageUrl, stock, sellerName]
...